feat: reorganize 404 error page; update import path and remove old component; add lisence headers
