+++ apps/wso2-public-wallet/src/components/home/recent-activities.js
@@ -1,3 +1,18 @@
+// Copyright (c) 2025 WSO2 LLC. (https://www.wso2.com).
+//
+// WSO2 LLC. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
 import React, { useEffect, useState } from 'react';
 import { RightOutlined, ArrowUpOutlined, ArrowDownOutlined, SmileOutlined } from "@ant-design/icons";
 import { Avatar, Spin, Card, Col, Row, Tag } from 'antd';
