+++ apps/wso2-public-wallet/src/components/home/wallet-address-copy.js
@@ -1,10 +1,24 @@
+// Copyright (c) 2025 WSO2 LLC. (https://www.wso2.com).
+//
+// WSO2 LLC. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
 import React, { useEffect, useState } from 'react';
 import { CopyOutlined, CheckOutlined } from '@ant-design/icons';
 import { CopyToClipboard } from 'react-copy-to-clipboard';
 import { Input, Button, message } from 'antd';
 import { COPIED, OK } from '../../constants/strings';
 import { defaultIconPrefixCls } from 'antd/es/config-provider';
-
 
 const WalletAddressCopy = (props) => {
 
@@ -31,7 +45,7 @@ const WalletAddressCopy = (props) => {
             setWalletAddressCopied(false);
         }, 2000)
     };
-    
+
 
     // --- return the wallet address copy component ---
     return (
