+++ apps/wso2-public-wallet/src/components/navbar/navbar.js
@@ -13,6 +13,7 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
+
 import React, { useState, useEffect } from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import { Layout, Menu, Row, Col, Tag, Switch } from 'antd';
