+++ apps/wso2-public-wallet/src/pages/confirm-token-send/confirm-token-send.js
@@ -1,9 +1,18 @@
-// Copyright (c) 2023, WSO2 LLC. (http://www.wso2.com). All Rights Reserved.
+// Copyright (c) 2025 WSO2 LLC. (https://www.wso2.com).
 //
-// This software is the property of WSO2 LLC. and its suppliers, if any.
-// Dissemination of any information or reproduction of any material contained
-// herein in any form is strictly forbidden, unless permitted by WSO2 expressly.
-// You may not alter or remove any copyright or other notice from copies of this content.
+// WSO2 LLC. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
 
 import React, { useState, useEffect } from "react";
 import Identicon from "identicon.js";
@@ -92,7 +101,7 @@ const ConfirmSendTokens = () => {
         try {
             await saveLocalDataAsync(STORAGE_KEYS.SENDING_AMOUNT, "");
             await saveLocalDataAsync(STORAGE_KEYS.SENDER_WALLET_ADDRESS, "");
-            
+
         } catch (error) {
             console.log(`${ERROR_RESETTING_TX_VALUES}: ${error}`);
         }
@@ -116,9 +125,9 @@ const ConfirmSendTokens = () => {
                     navigate("/");
                     setIsTransferLoading(false);
                 }, 3000);
-                
+
             }
-           
+
         } catch (error) {
             console.log("error while transferring token", error);
             await messageApi.error(ERROR, ERROR_TRANSFERRING_TOKEN);
